Improved:Account.java, CartItem.java, Order.java, OrderItem.java, OrderItemRowMapper.java, OrderRowMapper.java, Product.java, Shop.java, ShopDao.java, schema.sql Added: ShopRowMapper.java

+++ src/main/java/by/tms/onlinerclonec29onl/dao/mapper/OrderItemRowMapper.java
@@ -17,7 +17,7 @@ public class OrderItemRowMapper implements RowMapper<OrderItem> {
     @Override
     public OrderItem mapRow(ResultSet rs, int rowNum) throws SQLException {
         OrderItem orderItem = new OrderItem();
-        orderItem.setId(rs.getInt("main_orderitem_id"));
+        orderItem.setId(rs.getLong("main_orderitem_id"));
         orderItem.setProduct(productRowMapper.mapRow(rs, rowNum));
         orderItem.setQuantity(rs.getInt("quantity"));
         orderItem.setPrice(rs.getDouble("price"));
